+++ test/integration2/schema/misc.spec.js
@@ -483,24 +483,53 @@ describe('Schema (misc)', () => {
         });
 
         describe('uuid types - postgres', () => {
-          before(async () => {
-            await knex.schema.createTable('uuid_column_test', (table) => {
-              table.uuid('id', { primaryKey: true });
-            });
+          after(async () => {
+            if (isPgBased(knex)) {
+              await knex.schema.dropTable('uuid_column_test');
+            }
           });
 
-          after(async () => {
-            await knex.schema.dropTable('uuid_column_test');
+          it('creates a uuid column as primary using fluid syntax', async function () {
+            if (!isPgBased(knex)) {
+              return this.skip();
+            }
+            const table_name = 'uuid_column_test';
+            const expected_column = 'id';
+            const expected_type = 'uuid';
+
+            await knex.schema.dropTableIfExists(table_name);
+            await knex.schema.createTable(table_name, (table) => {
+              table.uuid('id').primary();
+            });
+
+            const cols = await knex.raw(
+              `select c.column_name, c.data_type
+                     from INFORMATION_SCHEMA.COLUMNS c
+                     join INFORMATION_SCHEMA.KEY_COLUMN_USAGE cu
+                     on (c.table_name = cu.table_name and c.column_name = cu.column_name)
+                     where c.table_name = ?
+                     and (cu.constraint_name like '%_pkey' or cu.constraint_name = 'primary')`,
+              table_name
+            );
+            const column_name = cols.rows[0].column_name;
+            const column_type = cols.rows[0].data_type;
+
+            expect(column_name).to.equal(expected_column);
+            expect(column_type).to.equal(expected_type);
           });
 
           it('#5211 - creates an uuid column as primary key', async function () {
             if (!isPgBased(knex)) {
               return this.skip();
             }
-
             const table_name = 'uuid_column_test';
             const expected_column = 'id';
             const expected_type = 'uuid';
+
+            await knex.schema.dropTableIfExists(table_name);
+            await knex.schema.createTable(table_name, (table) => {
+              table.uuid('id', { primaryKey: true });
+            });
 
             const cols = await knex.raw(
               `select c.column_name, c.data_type
@@ -862,8 +891,7 @@ describe('Schema (misc)', () => {
               tester(
                 ['pg', 'cockroachdb'],
                 [
-                  'create table "test_table_three" ("main" integer not null, "paragraph" text default \'Lorem ipsum Qui quis qui in.\', "metadata" json default \'{"a":10}\', "details" jsonb default \'{"b":{"d":20}}\')',
-                  'alter table "test_table_three" add constraint "test_table_three_pkey" primary key ("main")',
+                  'create table "test_table_three" ("main" integer not null, "paragraph" text default \'Lorem ipsum Qui quis qui in.\', "metadata" json default \'{"a":10}\', "details" jsonb default \'{"b":{"d":20}}\', constraint "test_table_three_pkey" primary key ("main"))',
                 ]
               );
               tester('pg-redshift', [
@@ -1626,6 +1654,43 @@ describe('Schema (misc)', () => {
               .where({
                 relname: 'phone_idx_2',
                 indisvalid: true,
+              })
+              .whereNotNull('indpred');
+            expect(results).to.not.be.empty;
+          });
+        });
+
+        describe('supports partial unique indexes - postgres, sqlite, and mssql', function () {
+          it('allows creating a unique index with predicate', async function () {
+            if (!(isPostgreSQL(knex) || isMssql(knex) || isSQLite(knex))) {
+              return this.skip();
+            }
+
+            await knex.schema.table('test_table_one', function (t) {
+              t.unique('email', {
+                indexName: 'email_idx',
+                predicate: knex.whereNotNull('email'),
+              });
+            });
+          });
+
+          it('actually stores the predicate in the Postgres server', async function () {
+            if (!isPostgreSQL(knex)) {
+              return this.skip();
+            }
+            await knex.schema.table('test_table_one', function (t) {
+              t.unique('email', {
+                indexName: 'email_idx_2',
+                predicate: knex.whereNotNull('email'),
+              });
+            });
+            const results = await knex
+              .from('pg_class')
+              .innerJoin('pg_index', 'pg_index.indexrelid', 'pg_class.oid')
+              .where({
+                relname: 'email_idx_2',
+                indisvalid: true,
+                indisunique: true,
               })
               .whereNotNull('indpred');
             expect(results).to.not.be.empty;